requirements.txt
# Core deep learning
torch>=2.0.1
torchvision>=0.15.2
torchaudio>=2.0.2

# Scientific computing
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.2.0
pandas>=2.0.0

# Computer vision
opencv-python>=4.7.0
Pillow>=9.5.0
albumentations>=1.3.0

# Face detection / landmarks
# RetinaFace via insightface (recommended) or standalone
insightface>=0.7.3
onnxruntime>=1.14.0

# Metrics
pingouin>=0.5.3          # ICC computation

# Visualisation
matplotlib>=3.7.0
seaborn>=0.12.0

# Configuration
PyYAML>=6.0
omegaconf>=2.3.0

# Logging & experiment tracking
tensorboard>=2.13.0
tqdm>=4.65.0
rich>=13.3.0

# Utilities
einops>=0.6.1            # Tensor rearrangement for transformers
timm>=0.9.2              # Pre-trained vision models

# Testing
pytest>=7.3.0
pytest-cov>=4.1.0

# Optional: FLIR thermal data reading
# flirpy>=0.4.0          # Uncomment if reading raw FLIR files
